+++ mypyc/irbuild/function.py
@@ -27,7 +27,7 @@ from mypyc.common import (
     SELF_NAME, ENV_ATTR_NAME, NEXT_LABEL_ATTR_NAME, LAMBDA_NAME, decorator_helper_name
 )
 from mypyc.sametype import is_same_method_signature
-from mypyc.genopsutil import concrete_arg_kind, is_constant, add_self_to_env
+from mypyc.irbuild.util import concrete_arg_kind, is_constant, add_self_to_env
 from mypyc.genopscontext import FuncInfo, GeneratorClass, ImplicitClass
 from mypyc.irbuild.statement import transform_try_except
 from mypyc.genops import IRBuilder
