+++ mypyc/irbuild/util.py
